requirements.txt
nengo
nengo-ocl
numpy
opencv-python
pillow
matplotlib
